repo-service/requirements.txt
fastapi[standard]
PyGithub
python-dotenv
